+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+* 1.0.0-alpha2 (@release_date@)
+
+ * added `$name` argument to `PackageManager::installPackage()`
+ * added `getName()` and `setName()` to `PackageMetadata`
+ * removed `setNew()` and `isNew()` from `PackageMetadata`
+ * changed `Package` to prefer the name of the metadata over the name set in the
+   package file
+
 * 1.0.0-alpha1 (2014-12-03)
 
  * first alpha release